requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
filelock==3.8.0
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
idna==3.4
mysql-connector-python==8.0.32
numpy==1.23.4
pi==0.1.2
Pillow==9.3.0
platformdirs==2.5.2
protobuf==3.20.3
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.5
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.15
virtualenv==20.16.5
